fix: Tidying up the SRR-related scraping and moving that logic to new collect.js
